refactor: move (un)-wrap fn to server-kit package
