+++ packages/core-kit/src/domains/key/index.ts
@@ -6,4 +6,3 @@
  */
 
 export * from './type';
-export * from './utils';
