+++ packages/server-core/src/http/controllers/root/jwks/handlers/read.ts
@@ -11,7 +11,7 @@ import { createPublicKey } from 'node:crypto';
 import type { Request, Response } from 'routup';
 import { send, useRequestParam } from 'routup';
 import { In } from 'typeorm';
-import { wrapPublicKeyPem } from '@authup/core-kit';
+import { wrapPublicKeyPem } from '@authup/server-kit';
 import { NotFoundError } from '@ebec/http';
 import { useDataSource } from 'typeorm-extension';
 import { KeyEntity } from '../../../../../domains';
