+++ packages/server-kit/src/crypto/key-pair/delete.ts
@@ -8,7 +8,7 @@
 import path from 'node:path';
 import fs from 'node:fs';
 import type { KeyPairOptions } from './type';
-import { buildKeyFileName, extendKeyPairOptions } from './utils';
+import { buildKeyFileName, extendKeyPairOptions } from './helpers';
 import { KeyPairKind } from './constants';
 
 export async function deleteKeyPair(context?: Partial<KeyPairOptions>) : Promise<void> {
