+++ packages/server-kit/src/crypto/key-pair/helpers/private-key.ts
