+++ packages/server-kit/src/crypto/key-pair/index.ts
@@ -12,4 +12,4 @@ export * from './load';
 export * from './module';
 export * from './save';
 export * from './type';
-export * from './utils';
+export * from './helpers';
